refactor: restructure `air.api` module and update imports for clarity
- Removed `air.api` module, redistributed components to relevant `air` submodules for a more concise structure.
- Updated imports across codebase and tests to align with the restructuring.
