+++ tests/test_applications.py
@@ -107,7 +107,7 @@ def test_air_404_response():
 def test_default_500_exception_handler():
     from starlette.requests import Request
 
-    from air.api.error_responses import default_500_exception_handler
+    from air.error_responses import default_500_exception_handler
 
     # Create a mock request and exception
     scope = {
@@ -137,7 +137,7 @@ def test_default_500_exception_handler():
 
 
 def test_injection_of_default_exception_handlers():
-    from air.api.error_responses import DEFAULT_EXCEPTION_HANDLERS
+    from air.error_responses import DEFAULT_EXCEPTION_HANDLERS
 
     def handler(request: air.Request, exc: Exception) -> air.AirResponse:
         return air.AirResponse()
